# Conflicts:
#	sandbox-backend/build.gradle
#	sandbox-backend/src/main/resources/db/changelog/db.changelog-master.yaml
#	sandbox-web/src/main/resources/application.properties

+++ sandbox-backend/src/main/java/com/exadel/team2/sandbox/service/impl/StatusHistoryServiceImpl.java
@@ -3,6 +3,7 @@ package com.exadel.team2.sandbox.service.impl;
 import com.exadel.team2.sandbox.dao.EmployeeDAO;
 import com.exadel.team2.sandbox.dao.StatusDAO;
 import com.exadel.team2.sandbox.dao.StatusHistoryDAO;
+import com.exadel.team2.sandbox.dao.rsql.CustomRsqlVisitor;
 import com.exadel.team2.sandbox.entity.EmployeeEntity;
 import com.exadel.team2.sandbox.entity.Status;
 import com.exadel.team2.sandbox.entity.StatusHistory;
@@ -11,9 +12,12 @@ import com.exadel.team2.sandbox.service.StatusHistoryService;
 import com.exadel.team2.sandbox.web.statushistory.CreateStatusHistoryDTO;
 import com.exadel.team2.sandbox.web.statushistory.ResponseStatusHistoryDTO;
 import com.exadel.team2.sandbox.web.statushistory.UpdateStatusHistoryDTO;
+import cz.jirutka.rsql.parser.RSQLParser;
+import cz.jirutka.rsql.parser.ast.Node;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -41,8 +45,14 @@ public class StatusHistoryServiceImpl implements StatusHistoryService {
     }
 
     @Override
-    public Page<ResponseStatusHistoryDTO> findAllPageable(Pageable pageable) {
-        return historyDAO.findAll(pageable).map(historyMapper::convertEntityToDto);
+    public Page<ResponseStatusHistoryDTO> findAllPageable(Pageable pageable, String query) {
+        if (query.isEmpty()) {
+            return historyDAO.findAll(pageable)
+                    .map(historyMapper::convertEntityToDto);
+        }
+        Node rootNode = new RSQLParser().parse(query);
+        Specification<StatusHistory> spec = rootNode.accept(new CustomRsqlVisitor<>());
+        return historyDAO.findAll(spec, pageable).map(historyMapper::convertEntityToDto);
 
     }
 
@@ -90,6 +100,8 @@ public class StatusHistoryServiceImpl implements StatusHistoryService {
 
     @Override
     public void deleteById(Long id) {
+        if (!historyDAO.existsById(id))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "StatusHistory id not found");
         historyDAO.deleteById(id);
     }
 }
